+++ docs/changelog.rst
@@ -1,5 +1,11 @@
 Changelog
 =========
+- :release:`1.1.0 <2023-07-24>`	
+- :feature:`-` Add support for a TreeLoad to be added to the model (#130)
+
+- :release:`1.0.1 <2023-06-14>`
+- :bug:`-` Fix: Updated content version of D-Stability to "2" (#124)
+
 - :release:`1.0.0 <2023-05-15>`
 - :feature:`-` Add support for D-Stability 2023.01 (#112). This release drops support for D-Stability 2022.01 and older.
 - :feature:`-` Removed support for Python 3.7, minimum supported Python version is now 3.8.